requirements.txt
Faker==24.4.0
azure-storage-file-datalake==12.14.0
python-dotenv==1.0.1
